backend/requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.0.3
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.2.0
Camelot==12.6.29
camelot-py==0.9.0
certifi==2022.9.24
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.7
comm==0.2.2
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cryptography==42.0.3
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dj-database-url==2.1.0
Django==5.1.5
django-bootstrap4==24.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Elixir==0.7.1
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
fastapi==0.109.2
fastjsonschema==2.19.1
ffs==0.0.8.2
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
gast==0.5.4
gcloud==0.17.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.63.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.21.0
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.4
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
jws==0.1.3
keras==3.5.0
kiwisolver==1.4.5
letter==0.5
libclang==18.1.1
looseversion==1.3.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.7
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==3.0.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
path==16.10.0
pathlib==1.0.1
pbr==6.0.0
pdfminer.six==20221105
pdfplumber==0.10.4
pdfrw==0.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
priority==2.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.21.9
prov==2.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.6.0
pylint==3.0.3
PyMySQL==1.1.1
pyparsing==3.0.9
pypdf==4.0.1
PyPDF2==3.0.1
pypdfium2==4.27.0
pypng==0.20220715.0
Pyrebase==3.0.27
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-jwt==2.0.1
pytz==2022.5
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
rdflib==7.0.0
referencing==0.34.0
reportlab==4.0.9
requests==2.28.1
requests-toolbelt==0.7.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.3
setuptools==69.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.3
stack-data==0.6.3
starlette==0.36.3
streamlit==1.33.0
svglib==1.5.1
tabula==1.0.5
tabulate==0.9.0
Tempita==0.5.2
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
traits==6.3.2
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2022.6
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.12
views==0.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.43.0
whitenoise==6.6.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.15
xlrd==0.7.1
xlwt==0.7.2
